Create app.ts
