+++ src/index.ts
@@ -1,35 +1,36 @@
-import express, { Request, Response, ErrorRequestHandler, Application} from "express";
-import dotenv from "dotenv";
-import cors from "cors";
+// import express, { Request, Response, ErrorRequestHandler, Application} from "express";
+// import dotenv from "dotenv";
+// import cors from "cors";
 import mongoose from "mongoose";
-import cookieParser from "cookie-parser";
-import authsRoutes from "./routes/authsRoutes";
-import companyRoutes from "./routes/companyRoutes";
-import officersRoutes from "./routes/officersRoutes";
-import sendersRoutes from "./routes/sendersRoutes";
-import receiversRoutes from "./routes/receiversRoutes";
-import ordersRoutes from "./routes/ordersRoutes";
+// import cookieParser from "cookie-parser";
+// import authsRoutes from "./routes/authsRoutes";
+// import companyRoutes from "./routes/companyRoutes";
+// import officersRoutes from "./routes/officersRoutes";
+// import sendersRoutes from "./routes/sendersRoutes";
+// import receiversRoutes from "./routes/receiversRoutes";
+// import ordersRoutes from "./routes/ordersRoutes";
+import app from "./app";
 
-dotenv.config();
-const app: Application = express();
+// dotenv.config();
+// const app: Application = express();
 const port: number = Number(process.env.PORT) || 5001;
 
-/**Setup cors */
-// const allowedOrigins: string[] = ['http://localhost:3000'];
+// /**Setup cors */
+// // const allowedOrigins: string[] = ['http://localhost:3000'];
 
-// const options: cors.CorsOptions = {
-//   credentials: true,
-//   origin: allowedOrigins,
-// };
+// // const options: cors.CorsOptions = {
+// //   credentials: true,
+// //   origin: allowedOrigins,
+// // };
 
-// app.use(cors(options));
-app.use(cors());
+// // app.use(cors(options));
+// app.use(cors());
 
-/**Middleware to parse incoming request */
-app.use(express.json());
+// /**Middleware to parse incoming request */
+// app.use(express.json());
 
-/**cookie parser */
-app.use(cookieParser());
+// /**cookie parser */
+// app.use(cookieParser());
 
 /**connect to database */
 mongoose.set('strictQuery', true);
@@ -55,29 +56,29 @@ mongoose.connection.on("connected", () => {
 
 
 
-app.get('/api/v1/', (req: Request, res: Response) => {
-  res.send("Hello World")
-});
-app.use("/api/v1/auths", authsRoutes);
-app.use("/api/v1/companies", companyRoutes);
-app.use("/api/v1/officers", officersRoutes);
-app.use("/api/v1/senders", sendersRoutes );
-app.use("/api/v1/receivers", receiversRoutes);
-app.use("/api/v1/orders", ordersRoutes);
+// app.get('/api/v1/', (req: Request, res: Response) => {
+//   res.send("Hello World")
+// });
+// app.use("/api/v1/auths", authsRoutes);
+// app.use("/api/v1/companies", companyRoutes);
+// app.use("/api/v1/officers", officersRoutes);
+// app.use("/api/v1/senders", sendersRoutes );
+// app.use("/api/v1/receivers", receiversRoutes);
+// app.use("/api/v1/orders", ordersRoutes);
 
 
-/**Error Handler */
-const errorHandler: ErrorRequestHandler = (error, req, res, next) => {
-  res.status(error.status || 500)
-  res.json({
-    success: false,
-    status: error.status || 500,
-    message: error.message || "Sorry! something went wrong"
-  })
-  next();
-}
+// /**Error Handler */
+// const errorHandler: ErrorRequestHandler = (error, req, res, next) => {
+//   res.status(error.status || 500)
+//   res.json({
+//     success: false,
+//     status: error.status || 500,
+//     message: error.message || "Sorry! something went wrong"
+//   })
+//   next();
+// }
 
-app.use(errorHandler);
+// app.use(errorHandler);
 
 app.listen(port, () => {
   connect();
